requirements.txt
streamlit
openai
pdfplumber
pytesseract
Pillow
pymupdf
